feat:refactoring components name
