+++ src/components/burger-ingredients/components/product-card/product-card.jsx
@@ -2,11 +2,11 @@
 import React from "react";
 import PropTypes from 'prop-types';
 import { CurrencyIcon, Counter } from '@ya.praktikum/react-developer-burger-ui-components';
-import ProductCardStyles from './ProductCard.module.css';
+import styles from './styles.module.css';
 
 function ProductCard({ name, image, price }) {
   return (
-    <div className={`${ProductCardStyles.card} pl-4 pr-4`}>
+    <div className={`${styles.card} pl-4 pr-4`}>
       <Counter count={1} size="default" />
       <img className="" src={image} alt={name} />
       <p className="text text_type_digits-default mt-1 mb-1">
