+++ src/components/burger-ingredients/components/product-list/product-list.jsx
@@ -4,17 +4,17 @@
 /* eslint-disable react/forbid-prop-types */
 import React from 'react';
 import PropTypes from 'prop-types';
-import ProductListStyles from './ProductList.module.css';
-import ProductCard from '../productCard/ProductCard';
+import styles from './styles.module.css';
+import ProductCard from '../product-card/product-card';
 
 function ProductList({ items, itemsType, onIngredientClick }) {
   return (
-    <div className={`${ProductListStyles.container} mb-10`}>
+    <div className={`${styles.container} mb-10`}>
       <h2 className="text text_type_main-medium">{itemsType.name}</h2>
-      <ul className={`${ProductListStyles.list} pr-2 pl-4`}>
+      <ul className={`${styles.list} pr-2 pl-4`}>
         {items.map((item) => (
           <li
-            className={ProductListStyles.item}
+            className={styles.item}
             key={item._id}
             onClick={() => onIngredientClick(item)}
           >
