+++ src/components/ingredient-details/components/nutrition-value.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import NutritionValueStyles from './NutritionValue.module.css';
+import styles from './styles.module.css';
 
 function NutritionValue({ text, value }) {
   return (
-    <li className={`${NutritionValueStyles.nutritionListItem}`}>
+    <li className={`${styles.nutritionListItem}`}>
       <p className="text text_type_main-default text_color_inactive mb-2">{text}</p>
       <span className="text text_type_digits-default text_color_inactive">{value}</span>
     </li>
